requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# API & Integrations
openai==0.28.1
sendgrid==6.11.0
httpx==0.25.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2